% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MER.R
\name{get_mer}
\alias{get_mer}
\title{get_MER}
\usage{
get_mer(path = here::here("Data"), download = F)
}
\arguments{
\item{path}{The path to the raw MSD file or the folder to save it in if download == T}

\item{download}{Whether to download a new MSD file, defaults to FALSE}
}
\description{
get_MER
}
