% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppmr.R
\name{get_ppmr}
\alias{get_ppmr}
\title{get_ppmr}
\usage{
get_ppmr(path = here::here("Data"), download = F)
}
\arguments{
\item{path}{The path to the raw PPMR file or the folder to save it in if download == T}

\item{download}{Whether to download a new PPMR file, defaults to FALSE}
}
\description{
get_ppmr
}
