% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xwalk.R
\name{get_xwalk}
\alias{get_xwalk}
\title{Download and munge crosswalk file from PSM}
\usage{
get_xwalk(path = crosswalk)
}
\arguments{
\item{path}{path where you want file stored, default = "Data/xwalk"}
}
\description{
Download and munge crosswalk file from PSM
}
