% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MER.R
\name{get_mer}
\alias{get_mer}
\title{get_MER}
\usage{
get_mer(filepath = mer_data, filename)
}
\arguments{
\item{filepath}{path of mer data. Default = mer_data}

\item{filename}{name of the MSD or genie export}
}
\description{
get_MER
}
