% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_fact_df.R
\name{sc_fact_df}
\alias{sc_fact_df}
\title{sc_fact_df}
\usage{
sc_fact_df(filepath = sc_fact)
}
\arguments{
\item{filepath}{where sc_fact data set was read in via \code{get_scfact}}
}
\description{
sc_fact_df
}
